[topo] reorganize tokens
