+++ topo/src/slot.rs
